profile/aiop/central/src/androidTest/java/org/im97mori/ble/profile/aiop/central/db/AutomationIOProfileBondedDatabaseHelperTest.java
package org.im97mori.ble.profile.aiop.central.db;

import androidx.test.core.app.ApplicationProvider;

import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

public class AutomationIOProfileBondedDatabaseHelperTest {

    @Test
    public void test_constructor_00001() {
        assertNotNull(AutomationIOProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_constructor_00002() {
        assertEquals(AutomationIOProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()), AutomationIOProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_getProfileName_00001() {
        assertEquals(AutomationIOProfileBondedDatabaseHelper.PROFILE_NAME, AutomationIOProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()).getProfileName());
    }
}
